pyproject.toml
[tool.poetry]
name = "repl_python3_beta_Drinks-Cabin"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^1.1.2"
pymongo = "^3.11.3"
werkzeug = "^1.0.1"
dnspython = "^2.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
